Cleaned up __init__s, made user, version, and zenodo files in system
