+++ src/sas/qtgui/MainWindow/UnitTesting/AboutBoxTest.py
@@ -8,6 +8,7 @@ from PyQt5 import QtCore
 from unittest.mock import MagicMock
 
 import sas.sasview
+import sas.system.version
 from sas import config
 from sas.system import web, legal
 
@@ -45,7 +46,7 @@ class AboutBoxTest(unittest.TestCase):
         """
         version = self.widget.lblVersion
         self.assertIsInstance(version, QtWidgets.QLabel)
-        self.assertEqual(str(version.text()), str(sas.__version__))
+        self.assertEqual(str(version.text()), str(sas.system.version.__version__))
 
     def testAbout(self):
         """
@@ -53,8 +54,6 @@ class AboutBoxTest(unittest.TestCase):
         """
         about = self.widget.lblAbout
         self.assertIsInstance(about, QtWidgets.QLabel)
-        # build version
-        self.assertIn(str(sas.sasview.__build__), about.text())
         # License
         self.assertIn(str(legal.copyright), about.text())
         # URLs
