+++ src/sas/qtgui/MainWindow/WelcomePanel.py
@@ -1,6 +1,7 @@
 from PyQt5 import QtWidgets
 
 import sas.sasview
+import sas.system.version
 
 from sas import config
 from sas.system import legal
@@ -14,10 +15,9 @@ class WelcomePanel(QtWidgets.QDialog, Ui_WelcomePanelUI):
 
         self.setWindowTitle("Welcome")
 
-        version = sas.sasview.__version__ # TODO: Make consistent with other version references
-        build = sas.sasview.__build__ # TODO: Make consistent with other build references
+        version = sas.system.version.__version__  # TODO: Make consistent with other version references
 
-        ver = "\nSasView %s\nBuild: %s\n%s" % (version, build, legal.copyright)
+        ver = "\nSasView %s\n%s" % (version, legal.copyright)
 
         self.lblVersion.setText(ver)
