--- src/sas/sasview/__init__.py
@@ -1,8 +1,0 @@
-from distutils.version import StrictVersion
-# The version number must follow StrictVersion rules as outlined
-# in http://epydoc.sourceforge.net/stdlib/distutils.version.StrictVersion-class.html
-__version__ = "5.0.5"
-StrictVersion(__version__)
-__DOI__ = "Zenodo, 10.5281/zenodo.6331344"
-__release_date__ = "2022"
-__build__ = "GIT_COMMIT"
