+++ src/sas/system/user.py
@@ -1,0 +1,24 @@
+import os
+
+_USER_DIR = None
+
+def make_user_dir():
+    """
+    Create the user directory ~/.sasview if it doesn't already exist.
+    """
+    path = os.path.join(os.path.expanduser("~"),'.sasview')
+    if not os.path.exists(path):
+        os.mkdir(path)
+    return path
+
+
+def get_user_dir():
+    """
+    The directory where the per-user configuration is stored.
+
+    Returns ~/.sasview, creating it if it does not already exist.
+    """
+    global _USER_DIR
+    if not _USER_DIR:
+        _USER_DIR = make_user_dir()
+    return _USER_DIR
