+++ src/sas/system/version.py
@@ -1,0 +1,7 @@
+from distutils.version import StrictVersion
+
+__version__ = "5.0.5"
+__release_date__ = "2022"
+
+
+StrictVersion(__version__)
