+++ src/sas/system/zenodo.py
@@ -1,0 +1,1 @@
+__DOI__ = "Zenodo, 10.5281/zenodo.6331344"
